Initial Commit: Project Setup
This initial commit marks the start of our project. In this commit, we have set up the basic project structure, including directories, configuration files.

Including:
- Added project directories.
- Initialized version control with Git.
- Added .gitignore to exclude unnecessary files from version control.
- Set up project configuration files.

This commit is the foundation on which we will build the rest of the project. From here, we'll start working on feature development, bug fixes, and further project enhancements.